Update internal data state format
This changes the format for the overall data state
from one where all of the columns from each view
get merged into the top-level of an object to one
where the top-level object has keys corresponding
to the view names and the columns are stored
under that key. With the addition of a class to the
schema that composes all of the environment and
analysis type templates, the this state
representation is directly validate-able.